what we got so far
