+++ distributors.cabal
@@ -31,14 +31,16 @@ library
       Control.Lens.Bifocal
       Control.Lens.Diopter
       Control.Lens.Grammar
+      Control.Lens.Grammar.BackusNaur
+      Control.Lens.Grammar.Kleene
+      Control.Lens.Grammar.Stream
+      Control.Lens.Grammar.Symbol
+      Control.Lens.Grammar.Token
       Control.Lens.Grate
       Control.Lens.Internal.Equator
       Control.Lens.Internal.NestedPrismTH
       Control.Lens.Monocle
       Control.Lens.PartialIso
-      Control.Lens.RegEx
-      Control.Lens.Stream
-      Control.Lens.Token
       Control.Lens.Wither
       Data.Profunctor.Distributor
       Data.Profunctor.Monadic
