+++ src/Control/Lens/Grammar/Kleene.hs
@@ -1,0 +1,10 @@
+module Control.Lens.Grammar.Kleene
+  ( KleeneStarAlgebra (..)
+  ) where
+
+class Monoid a => KleeneStarAlgebra a where
+  starK :: a -> a
+  plusK :: a -> a
+  optK :: a -> a
+  altK :: a -> a -> a
+  empK :: a
